+++ src/main/java/eu/siacs/conversations/ui/StartConversationActivity.java
@@ -15,6 +15,7 @@ import android.os.Build;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
+import android.support.design.widget.TextInputLayout;
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
@@ -405,14 +406,18 @@ public class StartConversationActivity extends XmppActivity implements XmppConne
 
 	protected void shareBookmarkUri(int position) {
 		Bookmark bookmark = (Bookmark) conferences.get(position);
+		shareAsChannel(this, bookmark.getJid().asBareJid().toEscapedString());
+	}
+
+	public static void shareAsChannel(final Context context, final String address) {
 		Intent shareIntent = new Intent();
 		shareIntent.setAction(Intent.ACTION_SEND);
-		shareIntent.putExtra(Intent.EXTRA_TEXT, "xmpp:" + bookmark.getJid().asBareJid().toEscapedString() + "?join");
+		shareIntent.putExtra(Intent.EXTRA_TEXT, "xmpp:" + address + "?join");
 		shareIntent.setType("text/plain");
 		try {
-			startActivity(Intent.createChooser(shareIntent, getText(R.string.share_uri_with)));
+			context.startActivity(Intent.createChooser(shareIntent, context.getText(R.string.share_uri_with)));
 		} catch (ActivityNotFoundException e) {
-			Toast.makeText(this, R.string.no_application_to_share_uri, Toast.LENGTH_SHORT).show();
+			Toast.makeText(context, R.string.no_application_to_share_uri, Toast.LENGTH_SHORT).show();
 		}
 	}
 
@@ -833,6 +838,7 @@ public class StartConversationActivity extends XmppActivity implements XmppConne
 				if (uri != null) {
 					Invite invite = new Invite(intent.getData(), intent.getBooleanExtra("scanned", false));
 					invite.account = intent.getStringExtra("account");
+					invite.forceDialog = intent.getBooleanExtra("force_dialog", false);
 					return invite.invite();
 				} else {
 					return false;
@@ -845,7 +851,7 @@ public class StartConversationActivity extends XmppActivity implements XmppConne
 		List<Contact> contacts = xmppConnectionService.findContacts(invite.getJid(), invite.account);
 		if (invite.isAction(XmppUri.ACTION_JOIN)) {
 			Conversation muc = xmppConnectionService.findFirstMuc(invite.getJid());
-			if (muc != null) {
+			if (muc != null && !invite.forceDialog) {
 				switchToConversationDoNotAppend(muc, invite.getBody());
 				return true;
 			} else {
@@ -1000,7 +1006,7 @@ public class StartConversationActivity extends XmppActivity implements XmppConne
 	}
 
 	@Override
-	public void onJoinDialogPositiveClick(Dialog dialog, Spinner spinner, AutoCompleteTextView jid, boolean isBookmarkChecked) {
+	public void onJoinDialogPositiveClick(Dialog dialog, Spinner spinner, TextInputLayout layout, AutoCompleteTextView jid, boolean isBookmarkChecked) {
 		if (!xmppConnectionServiceBound) {
 			return;
 		}
@@ -1008,17 +1014,26 @@ public class StartConversationActivity extends XmppActivity implements XmppConne
 		if (account == null) {
 			return;
 		}
-		final Jid conferenceJid;
+		final String input = jid.getText().toString();
+		Jid conferenceJid;
 		try {
-			conferenceJid = Jid.of(jid.getText().toString());
+			conferenceJid = Jid.of(input);
 		} catch (final IllegalArgumentException e) {
-			jid.setError(getString(R.string.invalid_jid));
-			return;
+			final XmppUri xmppUri = new XmppUri(input);
+			if (xmppUri.isJidValid() && xmppUri.isAction(XmppUri.ACTION_JOIN)) {
+				final Editable editable = jid.getEditableText();
+				editable.clear();
+				editable.append(xmppUri.getJid().toEscapedString());
+				conferenceJid = xmppUri.getJid();
+			} else {
+				layout.setError(getString(R.string.invalid_jid));
+				return;
+			}
 		}
 
 		if (isBookmarkChecked) {
 			if (account.hasBookmarkFor(conferenceJid)) {
-				jid.setError(getString(R.string.bookmark_already_exists));
+				layout.setError(getString(R.string.bookmark_already_exists));
 			} else {
 				final Bookmark bookmark = new Bookmark(account, conferenceJid.asBareJid());
 				bookmark.setAutojoin(getBooleanPreference("autojoin", R.bool.autojoin));
@@ -1273,6 +1288,8 @@ public class StartConversationActivity extends XmppActivity implements XmppConne
 	private class Invite extends XmppUri {
 
 		public String account;
+
+		public boolean forceDialog = false;
 
 		public Invite(final Uri uri) {
 			super(uri);